Fix non key item pop NMs
Recently added Ironclad NMs are popped with regular items only, not key
items. The logic for checking whether the player possessed all items
required to pop the tracked NM was flawed in that it used a ternary and
fell back to KI checking. This meant that if the player had a key item
with the same ID as the item required, EPT would display green as if the
player was ready to pop.